requirements.txt
APScheduler==3.10.4
attrs==23.2.0
Automat==22.10.0
backports.zoneinfo==0.2.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
environs==11.0.0
fake-useragent==1.5.1
feedparser==6.0.11
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
incremental==22.10.0
ip-area==0.0.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
mysql-connector-python==8.3.0
packaging==24.0
parsel==1.9.1
peewee==3.17.1
Protego==0.3.1
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
scrapyd==1.4.3
scrapyd-api==0.0.1
scrapyd-client==1.2.3
service-identity==24.1.0
session-request==0.0.4
sgmllib3k==1.0.0
six==1.16.0
spider-admin-pro==2.0.13
SQLAlchemy==2.0.29
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.11.0
tzlocal==5.2
ua-parser==0.18.0
uberegg==0.1.1
urllib3==2.2.1
user-agents==2.2.0
w3lib==2.1.2
Werkzeug==3.0.2
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
